pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "collmex-cli"
version = "0.1.0"
description = "LLM-gesteuertes FiBu & Controlling CLI fuer Collmex (deutsche Buchhaltungssoftware)"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Christopher Helm", email = "me@christopher-helm.com" },
]
keywords = [
    "collmex",
    "buchhaltung",
    "fibu",
    "controlling",
    "skr03",
    "datev",
    "accounting",
    "cli",
    "germany",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: Other/Proprietary License",
    "Natural Language :: German",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mkdocs-material>=9.5",
]

[project.urls]
Homepage = "https://github.com/atraining/collmex-cli"
Repository = "https://github.com/atraining/collmex-cli"
Issues = "https://github.com/atraining/collmex-cli/issues"

[project.scripts]
collmex = "collmex.cli:main"

[tool.setuptools.packages.find]
include = ["collmex*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not live'"
markers = [
    "live: Tests gegen echte Collmex-API (run with: pytest -m live)",
]

[tool.coverage.run]
source = ["collmex"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B904",   # raise ... from err (stylistic, touched later)
    "UP042",  # StrEnum migration (Python 3.11+, touched later)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["F841"]  # unused locals in tests are often intentional

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
